Task09 [Кадыков]

+++ task09/src/com/example/task09/Task09.java
@@ -8,16 +8,15 @@ public class Task09 {
         float a = 1.0f;
         float b = 3.0f;
         float c = 1.0e9f;
-        double x = (a / b - 1.0 / 3.0) * c;
-
+        double x = (a / b - (float) (1.0 / 3.0)) * c;
         return x;
     }
 
     public static void main(String[] args) {
         // Здесь вы можете вручную протестировать ваше решение
-        /*
+
         System.out.println(solution() == 0.0d);
-        */
+
     }
 
 }
